requirements.txt
bcrypt==4.1.2
dnspython==2.8.0
numpy==2.4.2
pandas==3.0.0
passlib==1.7.4
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
tzdata==2025.3
